DERBY-5840: Clean up compiler warnings introduced by using Java 5 language features
Make DataValueDescriptor.setBigDecimal() take a BigDecimal instead of
a Number. The argument is always null or a BigDecimal instance, but
its declared type had to be Number in order to compile on CDC, which
is no longer supported.

By making the declared type BigDecimal, we avoid an unchecked cast to
Comparable in NumberDataType's implementation of the method.

Also remove the classes that contain the CDC-specific implementation
of the DECIMAL data type.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1485076 13f79535-47bb-0310-9956-ffa450edef68