DERBY-6315: Improve test coverage of org.apache.derby.impl.io.InputStreamFile
This patch was contributed by Ahsan Shamsudeen (ahsan dot competition at gmail dot com)

This patch removes the unused method StorageFile.length().

The length method is removed from the StorageFile interface, as well as
from the InputStreamFile, JarDBFile, and CorruptFile implementations.

The method was not called anywhere, and removing it simplifies this area
of the Derby code.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1523334 13f79535-47bb-0310-9956-ffa450edef68