DERBY-534: Support use of the WHEN clause in CREATE TRIGGER statements
Allow references to transition variables and transition tables in the
WHEN clause. To support this, a new column WHENCLAUSETEXT is added to
the SYS.SYSTRIGGERS table, and a corresponding field is added to the
TriggerDescriptor class.

The logic that transforms triggered SQL statements to internal syntax
for accessing the transition variables and transition tables (via Java
method calls and VTIs) is reused on the WHEN clause text so that the
same transformation happens there.

Upgrade logic is added so that the new column in SYS.SYSTRIGGERS will
be created when a database is upgraded from an older version. The WHEN
clause is now disabled in the parser when running in soft upgrade
mode. An upgrade test case checks that the WHEN clause can only be
used in a hard-upgraded database, and that a reasonable error is
raised otherwise.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1526831 13f79535-47bb-0310-9956-ffa450edef68