DERBY-4923 update of a long row can fail with ERROR nospc: nospc.U exception.
This checkin fixes the problem repro'd by the included new test, which shows
an update of an existing row in db failing with a nosp.U exception. 
The problem was an off by 1 error in checking for enough space to update a row
on an overflow page to just an overflow pointer. The intent of the code is to 
always reserve enough space in every overflow row to allow for this update. 
In this case there was exactly enough space, but the code mistaken thought it 
needed 1 more byte.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1535413 13f79535-47bb-0310-9956-ffa450edef68