DERBY-6383(Update trigger defined on one column fires on update of other columns).
This regression is caused by DERBY-4874(Trigger does not recognize new size of VARCHAR column expanded with ALTER TABLE. It fails with ERROR 22001: A truncation error was
encountered trying to shrink VARCHAR)

The regression is for Statement level triggers. The statement trigger gets fired incorrectly for any column update rather than just the column specified in the UPDATE of column clause. The fix is going to ensure that SYSTRIGGERS has the correct list of columns on which trigger should be fired. Row level triggers work fine. The only workaround at this point is to drop and recreate the statement triggers. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1537593 13f79535-47bb-0310-9956-ffa450edef68