DERBY-6082: Removed re-prepares trigged by openOnServer_
Changed piggy-backing of prepare of preparedStatementForAutogeneratedKeys_ so that it doesn't need openOnServer_ 

Removed MaterialStatement.markClosedOnServer and implementation since it did not do anything.

Since this reduced ClientStatement.markClosedOnServer() to a cleanup-method for section_, it was removed and section_ was encapsulated as a private member where setSection is responsible for cleanup.

Patch file derby-6082a.diff. (+ review comment) 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1539540 13f79535-47bb-0310-9956-ffa450edef68