DERBY-6362: CHECK constraint uses wrong schema for unqualified routine invocations
Part 2: Fix offset of qualified table names. Their begin offset
pointed to the beginning of the table name token, but should point to
the beginning of the schema name token if they have one.

Also leave begin and end offset unspecified (-1 implied) for some
synthetic table name nodes that previously set the offsets to 0
explicitly. This makes it easier to tell whether they are synthetic
later.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1548128 13f79535-47bb-0310-9956-ffa450edef68