DERBY-5866: Triggers fire out of order
Make sure the creation timestamp of a newly created trigger is higher
than the timestamps of the existing triggers on the same table. If it
is not higher, the triggers may fire in the wrong order. If the
current timestamp is not higher than the timestamp of the previous
trigger defined on the table, the new trigger now gets a creation
timestamp that is one millisecond higher than the creation timestamp
of the previous trigger.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1555702 13f79535-47bb-0310-9956-ffa450edef68