DERBY-6408: EXISTS returns NULL instead of FALSE
Make sure EXISTS subqueries are wrapped in an IsNullNode to convert
NULL to a BOOLEAN. Before, only NOT EXISTS subqueries were wrapped.

Make RowResultSetNode.preprocess() invoke preprocess() on the result
columns instead of directly on the subquery nodes. Calling it directly
on the subquery nodes prevented the result columns from seeing the
added IsNullNode.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1573548 13f79535-47bb-0310-9956-ffa450edef68