DERBY-6516 Convert lang/nestedCommit.sql to JUnit NestedCommitTest
Patch derby-6516-gc to fix the intermittent error we saw on this
issue. The patch makes sure we close the result set for the statement
'values doConnCommitInt()' when executed inside a stored
procedure. Previously, it was executed by the stored procedure
"doConnStmtIntNoRS". We changed that to "doConnStmtInt" since this
statement does indeed have a result set.

The supporting procedure in Triggers#doConnStmt also had to be fixed a
bit, since this particular result set closes itself by performing a
commit inside the store procedure, which led the second ResultSet#next
to throw XLC16.

With the patch, I do not see the original error even when disabling the
explicit garbage collection (Java option -XX:+DisableExplicitGC).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1584638 13f79535-47bb-0310-9956-ffa450edef68