DERBY-6410: ClassCastException when launching derby from windows subst drive
Avoid calling Files.getFileStore() since it fails when called on a
file that lives on a subst drive. (Because of a bug in the JVM:
JDK-8034057.)

It is used for checking if a specific FileAttributeView is supported
on the file system. The same information can be found by checking if
Files.getFileAttributeView() returns null, which we already do, so
skip the failing call.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1585407 13f79535-47bb-0310-9956-ffa450edef68