DERBY-6566: Simplify handling of untyped nulls in CASE and NULLIF expressions
Make the parser represent untyped nulls the same way in CASE
expressions as in NULLIF expressions, so that ConditionalNode can
handle the two kinds of expressions uniformly.

Make the parser create a single ConditionalNode for each CASE
expression, instead of creating a tree of ConditionalNodes for a
single CASE expression.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1593898 13f79535-47bb-0310-9956-ffa450edef68