DERBY-6642 Make the JUnit result set assertion methods dump the actual rs when assertion fails
Patch derby-6642b. The patch works by always collecting columns and
rows as they are read. The method JDBC#addRsToReport takes these
collections and what's left of the result set and formats a dump of
the result set and adds it to the text of already existing assertion
failure. This is then (re)thrown.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1607069 13f79535-47bb-0310-9956-ffa450edef68