DERBY-6617 Silently swallowed SecurityExceptions may disable Derby features, including security features.
Patch derby-6617-2. With reference to the list of issues mentioned in the JIRA, the following
"silent swallows" are addressed with this patch:

* FileMonitor PBgetJVMProperty Catch java.lang.SecurityException 1 line 183
* FileMonitor PBinitialize Catch java.lang.SecurityException 1 line 120
* FileMonitor PBinitialize Catch java.lang.SecurityException 1 line 157

These are not addressed in code yet:

  BaseMonitor readApplicationProperties Catch java.lang.SecurityException 1 line 1360
  BaseMonitor runWithState Catch java.lang.SecurityException 0 line 280
  FileMonitor createDaemonGroup Catch java.lang.SecurityException 1 line 89

A new test (MissingPermissionsTest) has been added to verify that we
get errors in either derby.log or on the console's standard error
(depending on what's possible) if we see a SecurityException in the
former "silent" swallow location.

In the case of FileMonitor line 157, the boot fails so we show the
error on the console (i.e. standard error). In order for that to
happen, we added a flushing of the temporary in-memory error log,
cf. added call to dumpTempWriter if we fail BaseMonitor#initialize in
console.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1609717 13f79535-47bb-0310-9956-ffa450edef68