DERBY-6628 Intermittent error in TriggerGeneralTest
Patch derby-6628-2.

We decided, instead of filering the unexpected system table locks, to
retry the assert for the lock result set for a while to rule out slow,
lingering (re)compilation locks (by the index statistics daemon?).

If the locks of the system tables hang around for a minute (max
waiting time), something fishy is going on.

We'll run with this for a while, and if we don't see this issue again,
close the issue.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1621223 13f79535-47bb-0310-9956-ffa450edef68