DERBY-4057 Space is not reclaimed if transaction is rolled back
Added infrastructure called at insert abort time to queue post abort work.
After the abort work is queued in the case of aborted inserts to reclaim
space and if possible mark pages free, which then in turn allows them to
be used by subsequent work on the table.  This work queues this work
on heap tables when the aborted insert is the last row on a page, or
if the aborted insert is a row that contains a long column (a row that
is bigger than a page - usually a blob or clob), or a long row (a row
that spans multiple pages).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1641418 13f79535-47bb-0310-9956-ffa450edef68