DERBY-6810: Add regression tests for XXE vulnerability
This change adds the so-called "billion laughs" XML attack to the
XMLXXETest suite of XXE test cases. See:

    https://en.wikipedia.org/wiki/Billion_laughs

Note that even though this suite runs with no security manager, the default
limit of 64,000 entity expansions still applies, as documented in the Java
documentation here:

    http://docs.oracle.com/javase/1.5.0/docs/guide/xml/jaxp/JAXP-Compatibility_150.html#JAXP_security

And thus Derby, even with no security manager, is not vulnerable to the
billion laughs attack, at least since J2SE version 5.

Even though this test case does not demonstrate any error in Derby behavior,
so far as I can tell, it is still a worthwhile test to have in the suite.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1685313 13f79535-47bb-0310-9956-ffa450edef68