DERBY-6783: WHEN clause in CREATE TRIGGER for UPDATE is not working.
This patch was contributed by Abhinav Gupta (abhinavgupta2004 at gmail dot com)

The main challenge raised by this issue involved the analysis of which
columns are referenced by the trigger. Trigger statements can reference
columns in a number of ways, including the "AFTER UPDATE OF" clause, the
trigger's action statement, and the trigger's "WHEN" clause.

When the WHEN clause support was introduced, this added a new way in which
a TRIGGER could reference columns, and the compiler code which translated
the trigger's logic into specific column references at runtime wasn't
getting the right set of columns.

The intent of this patch is to successfully run sql queries in Derby that
trigger updates in a table, affecting only a certain rows of data that
is determined by the when clause. This patch tries to cover a wide range
of possible ways in which rows are conditionally updated after an event.

Before this patch such trigger commands failed to execute correctly and
left the data in an inconsistent state.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1688062 13f79535-47bb-0310-9956-ffa450edef68