DERBY-6824: Move ShutdownException into shared code area
For reasons similar to those described in DERBY-4669, the new code in
o.a.d.shared.common.error is designed to be able to be present in multiple
jar files (e.g., in both derby.jar and in derbyclient.jar), so it needs
to be unsealed to avoid sealing violations when that occurs.

Marked the package unsealed.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1690482 13f79535-47bb-0310-9956-ffa450edef68