DERBY-6803: Change Sqlca and SqlException to call MessageUtils
This patch was contributed by Abhinav Gupta (abhinavgupta2004 at gmail dot com)

This patch modifies the code in the client libraries
which analyzes the error or exception information in
the sqlerrmc message data, to use the new features of
the MessageUtils shared code to decode the message format.

In particular, the code can now locate and decode the
arguments as well as the message id, and therefore can
include those arguments when constructing a custom
Derby-specific subclass of SQLIntegrityConstraintViolationException.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1694492 13f79535-47bb-0310-9956-ffa450edef68