DERBY-6773: Derby throws plain SQLIntegrityConstraintViolationException
This patch includes contributions from Abhinav Gupta (abhinavgupta2004 at gmail dot com)

This patch enables use of the new Derby-specific subclass
of SQLIntegrityConstraintViolationException in both the embedded
and client-server configurations of the SQLExceptionFactory.

In addition, a number of new tests are added to verify that
the new exception is thrown and can be caught as we expect,
and that the accessor methods on the new exception allow
applications to determine which table name and constraint
name were responsible for the exception.

We still have a couple cleanup tasks before DERBY-6773 is
fully complete, but as of this patch the new behavior is
in effect.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1696690 13f79535-47bb-0310-9956-ffa450edef68