DERBY-5950: PlanExporter and SignatureChecker don't appear in derbyrun.
This patch was contributed by Yun Fang (fangyuncn at gmail dot com)

This patch externalizes the help for 'java -jar derbyrun.jar' to reflect that
the PlanExporter and SignatureChecker tools are available from derbyrun.

The patch also enables PlanExporter from derbyrun, as it was not previously
available via this launch technique.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1727887 13f79535-47bb-0310-9956-ffa450edef68