DERBY-6865: RestrictiveFilePermissionsTest fails on Windows
Move the contents of FilePermissionServiceImpl into FileUtil. It was
originally separated out of FileUtil because it used functionality
only available on Java 7 and higher, so it had to be loaded via
reflection. Since Java 8 is the minimum level on trunk, it could be
used directly from FileUtil without reflection.

The reason for the test failure was that a recent change caused
FilePermissionServiceImpl to be excluded from derby.jar, so that its
services were not used, and the database files ended up with unexpected
permissions.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1728681 13f79535-47bb-0310-9956-ffa450edef68