DERBY-6828: Network Server fails to start in Czech locale.
LocalizedResource.getTextMessage has special exception-handling behavior
for situations where the chosen ResourceBundle does not contain the
message key that we are trying to format.

The exception handler simply displays the message key, together with
any arguments to the message.

But some messages have no arguments, for example the message
DRDA_SecurityInstalled.I has no arguments, and so NetworkServerControl
passes a null set of arguments when displaying that message.

Therefore, LocalizedResource.getTextMessage needs to include a null
guard, so that its exception handler doesn't trip over the non-existent
arguments and crash with a NullPointerException.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1730254 13f79535-47bb-0310-9956-ffa450edef68